+++ src/main/java/com/manchui/domain/chat/entity/mongodb/ChatMessageType.java
@@ -1,4 +1,4 @@
-package com.manchui.domain.entity.mongodb;
+package com.manchui.domain.chat.entity.mongodb;
 
 public enum ChatMessageType {
