+++ src/main/java/com/manchui/domain/chat/entity/mysql/ChatRoomUser.java
@@ -1,5 +1,6 @@
-package com.manchui.domain.entity;
+package com.manchui.domain.chat.entity.mysql;
 
+import com.manchui.domain.user.entity.User;
 import com.manchui.global.entity.Timestamped;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
