+++ src/main/java/com/manchui/domain/chat/repository/ChatRoomRepository.java
@@ -1,6 +1,6 @@
-package com.manchui.domain.repository;
+package com.manchui.domain.chat.repository;
 
-import com.manchui.domain.entity.ChatRoom;
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
