+++ src/main/java/com/manchui/domain/chat/service/ChatMessageService.java
@@ -1,18 +1,18 @@
-package com.manchui.domain.service;
+package com.manchui.domain.chat.service;
 
-import com.manchui.domain.dto.CustomUserDetails;
-import com.manchui.domain.dto.chat.ChatMessageRequest;
-import com.manchui.domain.dto.chat.ChatMessageResponse;
-import com.manchui.domain.dto.chat.ChatMessageSliceResponse;
-import com.manchui.domain.entity.mongodb.ChatMessageType;
-import com.manchui.domain.entity.ChatRoom;
-import com.manchui.domain.entity.ChatRoomUser;
-import com.manchui.domain.entity.User;
-import com.manchui.domain.entity.mongodb.ChatMessage;
-import com.manchui.domain.repository.ChatRoomRepository;
-import com.manchui.domain.repository.ChatRoomUserRepository;
-import com.manchui.domain.repository.UserRepository;
-import com.manchui.domain.repository.mongodb.ChatMessageRepository;
+import com.manchui.domain.auth.dto.CustomUserDetails;
+import com.manchui.domain.chat.dto.ChatMessageRequest;
+import com.manchui.domain.chat.dto.ChatMessageResponse;
+import com.manchui.domain.chat.dto.ChatMessageSliceResponse;
+import com.manchui.domain.chat.entity.mongodb.ChatMessageType;
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
+import com.manchui.domain.chat.entity.mysql.ChatRoomUser;
+import com.manchui.domain.user.entity.User;
+import com.manchui.domain.chat.entity.mongodb.ChatMessage;
+import com.manchui.domain.chat.repository.ChatRoomRepository;
+import com.manchui.domain.chat.repository.ChatRoomUserRepository;
+import com.manchui.domain.user.repository.UserRepository;
+import com.manchui.domain.chat.repository.mongodb.ChatMessageRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
