+++ src/main/java/com/manchui/domain/chat/service/ChatRoomService.java
@@ -1,14 +1,20 @@
-package com.manchui.domain.service;
+package com.manchui.domain.chat.service;
 
-import com.manchui.domain.dto.CustomUserDetails;
-import com.manchui.domain.dto.UserInfo;
-import com.manchui.domain.dto.chat.ChatRoomListDetail;
-import com.manchui.domain.dto.chat.ChatRoomListResponse;
-import com.manchui.domain.dto.chat.ChatRoomUserListResponse;
+import com.manchui.domain.chat.entity.mysql.ChatRoom;
+import com.manchui.domain.chat.entity.mysql.ChatRoomUser;
+import com.manchui.domain.chat.repository.ChatRoomRepository;
+import com.manchui.domain.chat.repository.ChatRoomUserRepository;
+import com.manchui.domain.auth.dto.CustomUserDetails;
+import com.manchui.domain.user.dto.UserInfo;
+import com.manchui.domain.chat.dto.ChatRoomListDetail;
+import com.manchui.domain.chat.dto.ChatRoomListResponse;
+import com.manchui.domain.chat.dto.ChatRoomUserListResponse;
 import com.manchui.domain.entity.*;
-import com.manchui.domain.entity.mongodb.ChatMessage;
+import com.manchui.domain.chat.entity.mongodb.ChatMessage;
 import com.manchui.domain.repository.*;
-import com.manchui.domain.repository.mongodb.ChatMessageRepository;
+import com.manchui.domain.chat.repository.mongodb.ChatMessageRepository;
+import com.manchui.domain.user.entity.User;
+import com.manchui.domain.user.repository.UserRepository;
 import com.manchui.global.exception.CustomException;
 import com.manchui.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
